move Postgres datatype classes to a new package
